Memory: Long statt Integer

+++ webapp/src/main/java/ch/zbw/sysVentorySaas/App/computers/Computers.java
@@ -1,7 +1,7 @@
 //
 // Diese Datei wurde mit der JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.8-b130911.1802 generiert 
 // Siehe <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Änderungen an dieser Datei gehen bei einer Neukompilierung des Quellschemas verloren. 
+// ï¿½nderungen an dieser Datei gehen bei einer Neukompilierung des Quellschemas verloren. 
 // Generiert: 2016.06.23 um 10:50:56 PM CEST 
 //
 
@@ -19,7 +19,7 @@ import javax.xml.bind.annotation.XmlType;
 
 
 /**
- * <p>Java-Klasse für anonymous complex type.
+ * <p>Java-Klasse fï¿½r anonymous complex type.
  * 
  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
  * 
@@ -413,7 +413,7 @@ public class Computers {
 
 
     /**
-     * <p>Java-Klasse für anonymous complex type.
+     * <p>Java-Klasse fï¿½r anonymous complex type.
      * 
      * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
      * 
@@ -899,7 +899,7 @@ public class Computers {
 
 
         /**
-         * <p>Java-Klasse für anonymous complex type.
+         * <p>Java-Klasse fï¿½r anonymous complex type.
          * 
          * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
          * 
@@ -1020,7 +1020,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -1261,7 +1261,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -1312,7 +1312,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -1363,7 +1363,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -1414,7 +1414,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -1465,7 +1465,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -1516,7 +1516,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -1571,7 +1571,7 @@ public class Computers {
 
 
         /**
-         * <p>Java-Klasse für anonymous complex type.
+         * <p>Java-Klasse fï¿½r anonymous complex type.
          * 
          * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
          * 
@@ -1688,7 +1688,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -1754,7 +1754,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -1807,7 +1807,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -1873,7 +1873,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -1928,7 +1928,7 @@ public class Computers {
 
 
         /**
-         * <p>Java-Klasse für anonymous complex type.
+         * <p>Java-Klasse fï¿½r anonymous complex type.
          * 
          * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
          * 
@@ -2013,7 +2013,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -2110,7 +2110,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -2161,7 +2161,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -2216,7 +2216,7 @@ public class Computers {
 
 
         /**
-         * <p>Java-Klasse für anonymous complex type.
+         * <p>Java-Klasse fï¿½r anonymous complex type.
          * 
          * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
          * 
@@ -2314,7 +2314,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -2483,7 +2483,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -2534,7 +2534,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -2585,7 +2585,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -2636,7 +2636,7 @@ public class Computers {
 
 
                 /**
-                 * <p>Java-Klasse für anonymous complex type.
+                 * <p>Java-Klasse fï¿½r anonymous complex type.
                  * 
                  * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
                  * 
@@ -2691,7 +2691,7 @@ public class Computers {
 
 
         /**
-         * <p>Java-Klasse für anonymous complex type.
+         * <p>Java-Klasse fï¿½r anonymous complex type.
          * 
          * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
          * 
@@ -2757,7 +2757,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -2810,7 +2810,7 @@ public class Computers {
 
 
         /**
-         * <p>Java-Klasse für anonymous complex type.
+         * <p>Java-Klasse fï¿½r anonymous complex type.
          * 
          * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
          * 
@@ -3051,7 +3051,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -3102,7 +3102,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -3153,7 +3153,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -3204,7 +3204,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -3225,7 +3225,7 @@ public class Computers {
             public static class Memory {
 
                 @XmlAttribute(name = "value")
-                protected Integer value;
+                protected Long value;
 
                 /**
                  * Ruft den Wert der value-Eigenschaft ab.
@@ -3235,7 +3235,7 @@ public class Computers {
                  *     {@link Integer }
                  *     
                  */
-                public Integer getValue() {
+                public Long getValue() {
                     return value;
                 }
 
@@ -3247,7 +3247,7 @@ public class Computers {
                  *     {@link Integer }
                  *     
                  */
-                public void setValue(Integer value) {
+                public void setValue(Long value) {
                     this.value = value;
                 }
 
@@ -3255,7 +3255,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              * 
@@ -3306,7 +3306,7 @@ public class Computers {
 
 
             /**
-             * <p>Java-Klasse für anonymous complex type.
+             * <p>Java-Klasse fï¿½r anonymous complex type.
              * 
              * <p>Das folgende Schemafragment gibt den erwarteten Content an, der in dieser Klasse enthalten ist.
              *
